auth

+++ src/classes/Comment.php
@@ -2,6 +2,15 @@
 
 class Comment extends Mapper
 {
+    private function getCommentAuthorByCommentId($commentID)
+    {
+        $statement = $this->db->prepare("SELECT createdBy FROM comments WHERE commentID = :commentID");
+        $statement->execute([
+            'commentID' => $commentID
+        ]);
+        return $statement->fetch(PDO::FETCH_COLUMN);
+    }
+
     public function createComment($entryID, $content)
     {
         $statement = $this->db->prepare("INSERT INTO comments(entryID, content, createdBy, createdAt) VALUES (:entryID, :content, :createdBy, NOW())");
@@ -14,10 +23,18 @@ class Comment extends Mapper
 
     public function deleteComment($commentID)
     {
-        $statement = $this->db->prepare("DELETE FROM comments WHERE commentID = :commentID");
-        $statement->execute([
-            'commentID' => $commentID
-        ]);
+        $createdBy = $this-> getCommentAuthorByCommentId($commentID);
+
+        if($createdBy === $_SESSION['userID']) {
+            $statement = $this->db->prepare("DELETE FROM comments WHERE commentID = :commentID");
+            $statement->execute([
+                'commentID' => $commentID
+            ]);
+
+            return true;
+        } else {
+            return false;
+        }
         
     }
     public function updateComment($commentID, $content) // $content will come from a variable that comes from JS
